output route_pattern_id to link with frequency data
calculate a field on the source table to join the geometry/network data back to it once it is output. then map that fiel to the Name field in the network output. this also meant moving some of the regex that did this stuff out of the functions, which seems better.