clarify dataframes in environment
there were a lot of dataframes with a lot of different names, many of which were just temporary. i moved temporary dataframes into functions preceded with the name df_ and then promoted dataframes that are key inputs and outputs to those functions to names that represent their ontology (e.g. PM_Peak_Bus_Routes is renamed pm_stops and Weekday_PM_Peak_High_Frequency_Bus_Service is renamed to pm_routes). some of the naming could probably be better but this should help with any potential debugging in any case