bad etl refactor
this was a well-meant but bad example of a refactor to the etl functions while testing out the scripts on another machine. the goal here was to break apart all the stuff thats going on in the combination of multiple csvs. see the commit to follow for a better version--where we just depend on a bunch of much better written etl functions in an ropensci package