removed arrival_time sorting for tables without arrival time
there was a legacy sort by arrival time (and other variables) that wasn't doing any work in the output. it was probably useful for checking on results visually but in this case it throws error when the value is empty. the changes here in r511 should be checked against outputs of the full-table method, but this was necessary in order to get the simpler versions of a few functions to work