clean up unused files; make r511.R match master
deleted files that aren't necessary for the output of high frequency bus geometries and their headway and stop calculations. these may be useufl for other things but will probably just be confusing for now if we need to come back to this later. also, master and r511.r have diverged a bit but this shuold supersede master as we've now incorporated changes on 8a2ce450af213707bbc6d61dbd035363b40f058c