refactor geometry grabbing and route id
to resolve https://github.com/MetropolitanTransportationCommission/RegionalTransitDatabase/issues/41 and https://github.com/MetropolitanTransportationCommission/RegionalTransitDatabase/issues/31 as discussed in https://github.com/MetropolitanTransportationCommission/RegionalTransitDatabase/issues/43. it makes sense to just solve these together because the route dataframes are being passed into the geometry grabbing operation. it would be better to be doing less of that--just first identifying routes that are high frequency and then getting the geometries for them. previously, we were putting stats on the headways in each direction, depending on the direction, and then appending those to geometries. this meant passing attributes around in weird ways in order to append them to spatial tables. but since all that matters is grabbing the right routes, based on the right route id, it would be easier to just ID the routes first and then grab the geometries.