write out routes for historical gtfs
using a csv of historical gtfs data for the bay area and gtfsr, write geopackages out to disk for each provider for each year and source. also, begin script to merge all of the output data for each year, which is the desired end result. it weas easier to debug the processing for this by provider, so just processing provider by provider before merge.