cache files fetched from 511
now that we have a list of cached transit data, we can use it to keep track of any new transit data we fetch. in order to process updated routes we cache the data to s3 first, then add it to the list of transit data we collect. this way we can just process feeds from s3 again and keep a record of the data that we processed.